requirements.txt
fastapi
uvicorn[standard]
gunicorn
jinja2
python-dotenv
sqlalchemy
psycopg2-binary
numpy
scipy
python-multipart
matplotlib